requirements.txt
aiogram==3.6.0
aiohttp==3.9.5
python-dotenv==1.0.1
requests==2.32.3
aiofiles~=23.2.1
PyJWT~=2.8.0
cryptography~=42.0.8
redis~=5.0.6
pydub~=0.25.1